No more reboot required, the fix will take effect instantly!
1. Hot fix.
2. More accurate permission of sysclusbfix.sleep.
3. Added uninstall function for fixusb.sh

+++ fixUSB.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 #  fixUSB.sh
-#  
+#
 #
 #  Created by syscl/lighting/Yating Zhou on 16/3/18.
 #
@@ -50,12 +50,20 @@ OFF="\033[m"
 REPO=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
 #
+# Define vars.
+#
+gArgv=""
+gDebug=1
+
+#
 # Path and filename setup.
 #
-gInstallDameon="/usr/local/sbin"
+gInstall_Repo="/usr/local/sbin/"
 gFrom="${REPO}/tools"
 gConfig="/tmp/de.bernhard-baehr.sleepwatcher.plist"
 gUSBSleepScript="/tmp/sysclusbfix.sleep"
+to_Plist="/Library/LaunchDaemons/com.syscl.externalfix.sleepwatcher.plist"
+to_shell="/etc/sysclusbfix.sleep"
 
 #
 #--------------------------------------------------------------------------------
@@ -84,6 +92,12 @@ function _PRINT_MSG()
                   then
                     local message=$(echo $message | sed -e 's/.*NOTE://')
                     echo "[ ${RED}NOTE${OFF} ] ${message}."
+                  else
+                    if [[ $message =~ 'DEBUG' ]];
+                      then
+                        local message=$(echo $message | sed -e 's/.*DEBUG://')
+                        echo "[${BLUE}DEBLOG${OFF}] ${message}."
+                    fi
                 fi
             fi
         fi
@@ -94,22 +108,31 @@ function _PRINT_MSG()
 #--------------------------------------------------------------------------------
 #
 
-function tidy_execute()
+function _tidy_exec()
 {
-    #
-    # Make the output clear.
-    #
-    $1 >/tmp/report 2>&1 && RETURN_VAL=0 || RETURN_VAL=1
-
-    if [ "${RETURN_VAL}" == 0 ];
+    if [ $gDebug -eq 0 ];
       then
-        _PRINT_MSG "OK: $2"
+        #
+        # Using debug mode to output all the details.
+        #
+        _PRINT_MSG "DEBUG: $2"
+        $1
       else
-        _PRINT_MSG "FAILED: $2"
-        cat /tmp/report
-    fi
+        #
+        # Make the output clear.
+        #
+        $1 >/tmp/report 2>&1 && RETURN_VAL=0 || RETURN_VAL=1
 
-    rm /tmp/report &> /dev/null
+        if [ "${RETURN_VAL}" == 0 ];
+          then
+            _PRINT_MSG "OK: $2"
+          else
+            _PRINT_MSG "FAILED: $2"
+            cat /tmp/report
+        fi
+
+        rm /tmp/report &> /dev/null
+    fi
 }
 
 #
@@ -118,10 +141,8 @@ function tidy_execute()
 
 function _printConfig()
 {
-    if [ -f ${gConfig} ];
-      then
-        rm ${gConfig}
-    fi
+
+    _del ${gConfig}
 
     echo '<?xml version="1.0" encoding="UTF-8"?>'                                                                                                           > "$gConfig"
     echo '<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">'                                          >> "$gConfig"
@@ -130,7 +151,7 @@ function _printConfig()
     echo '	<key>KeepAlive</key>'                                                                                                                          >> "$gConfig"
     echo '	<true/>'                                                                                                                                       >> "$gConfig"
     echo '	<key>Label</key>'                                                                                                                              >> "$gConfig"
-    echo '	<string>de.bernhard-baehr.sleepwatcher</string>'                                                                                               >> "$gConfig"
+    echo '	<string>com.syscl.externalfix.sleepwatcher</string>'                                                                                               >> "$gConfig"
     echo '	<key>ProgramArguments</key>'                                                                                                                   >> "$gConfig"
     echo '	<array>'                                                                                                                                       >> "$gConfig"
     echo '		<string>/usr/local/sbin/sleepwatcher</string>'                                                                                             >> "$gConfig"
@@ -149,10 +170,10 @@ function _printConfig()
 
 function _createUSB_Sleep_Script()
 {
-    if [ -f ${gUSBSleepScript} ];
-      then
-        rm ${gUSBSleepScript}
-    fi
+    #
+    # Remove previous script.
+    #
+    _del ${gUSBSleepScript}
 
     echo '#!/bin/sh'                                                                                                                                         > "$gUSBSleepScript"
     echo '#'                                                                                                                                                >> "$gUSBSleepScript"
@@ -168,7 +189,94 @@ function _createUSB_Sleep_Script()
     echo '# Added unmount Disk for "OS X" (c) syscl/lighting/Yating Zhou.'                                                                                  >> "$gUSBSleepScript"
     echo '#'                                                                                                                                                >> "$gUSBSleepScript"
     echo ''                                                                                                                                                 >> "$gUSBSleepScript"
-    echo 'diskutil list | grep -i "External" | sed -e "s| (external, physical):||" | xargs -I {} diskutil eject {}'                                   >> "$gUSBSleepScript"
+    echo 'diskutil list | grep -i "External" | sed -e "s| (external, physical):||" | xargs -I {} diskutil eject {}'                                         >> "$gUSBSleepScript"
+}
+
+#
+#--------------------------------------------------------------------------------
+#
+
+function _uinstall()
+{
+    _tidy_exec "sudo launchctl unload ${to_Plist}" "Unload ${to_Plist}"
+    _del /Library/LaunchDaemons/de.bernhard-baehr.sleepwatcher.plist
+    _del ${gInstall_Repo}
+    _del ${to_Plist}
+    _del ${to_shell}
+}
+
+#
+#--------------------------------------------------------------------------------
+#
+
+function _del()
+{
+    local target_file=$1
+
+    if [ -d ${target_file} ];
+      then
+        _tidy_exec "sudo rm -R ${target_file}" "Remove ${target_file}"
+      else
+        if [ -f ${target_file} ];
+          then
+            _tidy_exec "sudo rm ${target_file}" "Remove ${target_file}"
+        fi
+    fi
+}
+
+#
+#--------------------------------------------------------------------------------
+#
+
+function _touch()
+{
+    local target_file=$1
+
+    if [ ! -d ${target_file} ];
+      then
+        _tidy_exec "sudo mkdir ${target_file}" "Create ${target_file}"
+    fi
+}
+
+#
+#--------------------------------------------------------------------------------
+#
+
+function _install()
+{
+    #
+    # Generate configuration file of sleepwatcher launch demon.
+    #
+    _PRINT_MSG "--->: Generating configuration file of sleepwatcher launch daemon..."
+    _tidy_exec "_printConfig" "Generate configuration file of sleepwatcher launch daemon"
+
+    #
+    # Generate script to unmount external devices before sleep (c) syscl/lighting/Yating Zhou.
+    #
+    _PRINT_MSG "--->: Generating script to unmount external devices before sleep (c) syscl/lighting/Yating Zhou..."
+    _tidy_exec "_createUSB_Sleep_Script" "Generating script to unmount external devices before sleep (c) syscl/lighting/Yating Zhou"
+
+    #
+    # Install sleepwatcher daemon.
+    #
+    _PRINT_MSG "--->: Installing external devices sleep patch..."
+    _touch "${gInstall_Repo}"
+    _tidy_exec "sudo cp "${gFrom}/sleepwatcher" "${gInstall_Repo}"" "Install sleepwatcher daemon"
+    _tidy_exec "sudo cp "${gConfig}" "${to_Plist}"" "Install configuration of sleepwatcher daemon"
+    _tidy_exec "sudo cp "${gUSBSleepScript}" "${to_shell}"" "Install sleepwatcher script"
+    _tidy_exec "sudo chmod 744 ${to_shell}" "Fix the permissions of the file(s)"
+    _tidy_exec "sudo launchctl load ${to_Plist}" "Trigger startup service of syscl.usb.fix"
+
+    #
+    # Clean up.
+    #
+    _tidy_exec "rm $gConfig $gUSBSleepScript" "Clean up"
+
+    #
+    # Finish. Request reboot.
+    #
+    _PRINT_MSG "NOTE: DONE! Sleep to see change"
+    _PRINT_MSG "NOTE: Feel free to post issue on https://github.com/syscl/Fix-usb-sleep"
 }
 
 #
@@ -178,40 +286,43 @@ function _createUSB_Sleep_Script()
 function main()
 {
     #
-    # Generate configuration file of sleepwatcher launch demon.
+    # Get argument.
     #
-    _PRINT_MSG "--->: Generating configuration file of sleepwatcher launch daemon..."
-    tidy_execute "_printConfig" "Generate configuration file of sleepwatcher launch daemon"
+    gArgv=$(echo "$@" | tr '[:lower:]' '[:upper:]')
+    if [[ "$gArgv" == *"-D"* || "$gArgv" == *"-DEBUG"* ]];
+      then
+        #
+        # Yes, we do need debug mode.
+        #
+        _PRINT_MSG "NOTE: Use ${BLUE}DEBUG${OFF} mode"
+        gDebug=0
+      else
+        #
+        # No, we need a clean output style.
+        #
+        gDebug=1
+    fi
 
     #
-    # Generate script to unmount external devices before sleep (c) syscl/lighting/Yating Zhou.
+    # Detect which progress to execute.
     #
-    _PRINT_MSG "--->: Generating script to unmount external devices before sleep (c) syscl/lighting/Yating Zhou..."
-    tidy_execute "_createUSB_Sleep_Script" "Generating script to unmount external devices before sleep (c) syscl/lighting/Yating Zhou"
-
-    #
-    # Install sleepwatcher daemon.
-    #
-    _PRINT_MSG "--->: Install sleepwatcher daemon..."
-    tidy_execute "sudo cp "${gFrom}/sleepwatcher" "${gInstallDameon}"" "Install sleepwatcher daemon"
-    tidy_execute "sudo cp "${gConfig}" "/Library/LaunchDaemons"" "Install configuration of sleepwatcher daemon"
-    tidy_execute "sudo cp "${gUSBSleepScript}" "/etc"" "Install sleepwatcher script"
-    tidy_execute "sudo chmod +x /etc/sysclusbfix.sleep" "Change the permissions of the script (add +x) so that it can be run before sleep"
-
-    #
-    # Clean up.
-    #
-    tidy_execute "rm $gConfig $gUSBSleepScript" "Clean up"
-
-    #
-    # Finish. Request reboot.
-    #
-    _PRINT_MSG "NOTE: DONE! Please reboot to see change! If you have any questions or bugs to report, post issue on https://github.com/syscl/Fix-usb-sleep. THX :-)"
+    if [[ "$gArgv" == *"-U"* ]];
+      then
+        #
+        # Uninstall.
+        #
+        _uinstall
+      else
+        #
+        # Install.
+        #
+        _install
+    fi
 }
 
 #==================================== START =====================================
 
-main
+main "$@"
 
 #================================================================================
